Refactor process sandbox module boundaries
