+++ crates/ironclaw_host_runtime/src/services.rs
@@ -6,6 +6,8 @@
 //! facade. Authorization, run-state transitions, approval leases, process
 //! lifecycle, and runtime execution semantics remain in their owning crates.
 
+mod process_executor;
+
 use std::sync::{Arc, Mutex};
 
 use async_trait::async_trait;
@@ -27,9 +29,8 @@ use ironclaw_filesystem::LibSqlRootFilesystem;
 use ironclaw_filesystem::PostgresRootFilesystem;
 use ironclaw_filesystem::{LocalFilesystem, RootFilesystem, ScopedFilesystem};
 use ironclaw_host_api::{
-    CapabilityDispatchRequest, CapabilityDispatcher, CapabilityId, DispatchError,
-    ResourceReservationId, ResourceScope, ResourceUsage, RuntimeDispatchErrorKind,
-    RuntimeHttpEgress, RuntimeKind,
+    CapabilityDispatcher, CapabilityId, ResourceReservationId, ResourceScope, ResourceUsage,
+    RuntimeDispatchErrorKind, RuntimeHttpEgress, RuntimeKind,
     runtime_policy::{
         DeploymentMode, EffectiveRuntimePolicy, FilesystemBackendKind, NetworkMode,
         ProcessBackendKind, RuntimeProfile, SecretMode,
@@ -39,8 +40,7 @@ use ironclaw_mcp::{McpError, McpExecutionRequest, McpExecutor, McpInvocation};
 use ironclaw_network::NetworkHttpEgress;
 use ironclaw_processes::{
     BackgroundFailureStage, InMemoryProcessResultStore, InMemoryProcessStore,
-    ProcessExecutionError, ProcessExecutionRequest, ProcessExecutionResult, ProcessExecutor,
-    ProcessManager, ProcessResultStore, ProcessServices, ProcessStore,
+    ProcessExecutor, ProcessManager, ProcessResultStore, ProcessServices, ProcessStore,
 };
 use ironclaw_reborn_event_store::{
     RebornEventStoreConfig, RebornEventStoreError, RebornEventStores, RebornProfile,
@@ -80,6 +80,7 @@ use crate::{
     RuntimeBackendHealth, RuntimeProcessPort, TenantSandboxProcessPort, TurnRunExecutor,
     TurnRunScheduler, TurnRunSchedulerConfig, plan_capability,
 };
+use process_executor::{ProcessExecutorRouter, RuntimeDispatchProcessExecutor};
 
 type SharedRuntimeHttpEgress = Arc<Mutex<Option<Arc<dyn RuntimeHttpEgress>>>>;
 
@@ -321,10 +322,17 @@ where
     /// stores, cancellation registry, result store, and runtime health graph.
     fn build_host_runtime(&self) -> DefaultHostRuntime {
         let dispatcher: Arc<dyn CapabilityDispatcher> = Arc::new(self.runtime_dispatcher());
-        let process_executor = Arc::new(ProcessExecutorRouter::new(
-            Arc::new(RuntimeDispatchProcessExecutor::new(Arc::clone(&dispatcher))),
-            self.process_sandbox_executor.clone(),
-        ));
+        let process_executor = Arc::new(
+            ProcessExecutorRouter::new(Arc::new(RuntimeDispatchProcessExecutor::new(Arc::clone(
+                &dispatcher,
+            ))))
+            .with_optional_route(
+                RuntimeKind::System,
+                ironclaw_process_sandbox::PROCESS_SANDBOX_CAPABILITY_ID,
+                self.process_sandbox_executor.clone(),
+                "missing_process_sandbox_executor",
+            ),
+        );
         let lifecycle_process_store = Arc::clone(&self.process_lifecycle_store);
         let process_store: Arc<dyn ProcessStore> = lifecycle_process_store.clone();
         let result_failure_cleanup_store = Arc::clone(&lifecycle_process_store);
@@ -563,102 +571,6 @@ impl RuntimeBackendHealth for RegisteredRuntimeHealth {
             .collect::<Vec<_>>();
         normalize_runtime_kinds(&mut missing);
         Ok(missing)
-    }
-}
-
-#[derive(Clone)]
-struct RuntimeDispatchProcessExecutor {
-    dispatcher: Arc<dyn CapabilityDispatcher>,
-}
-
-impl RuntimeDispatchProcessExecutor {
-    pub(crate) fn new(dispatcher: Arc<dyn CapabilityDispatcher>) -> Self {
-        Self { dispatcher }
-    }
-}
-
-#[derive(Clone)]
-struct ProcessExecutorRouter {
-    dispatch_executor: Arc<dyn ProcessExecutor>,
-    process_sandbox_executor: Option<Arc<dyn ProcessExecutor>>,
-}
-
-impl ProcessExecutorRouter {
-    fn new(
-        dispatch_executor: Arc<dyn ProcessExecutor>,
-        process_sandbox_executor: Option<Arc<dyn ProcessExecutor>>,
-    ) -> Self {
-        Self {
-            dispatch_executor,
-            process_sandbox_executor,
-        }
-    }
-}
-
-#[async_trait]
-impl ProcessExecutor for ProcessExecutorRouter {
-    async fn execute(
-        &self,
-        request: ProcessExecutionRequest,
-    ) -> Result<ProcessExecutionResult, ProcessExecutionError> {
-        if is_process_sandbox_process(&request) {
-            let Some(executor) = &self.process_sandbox_executor else {
-                return Err(ProcessExecutionError::new(
-                    "missing_process_sandbox_executor",
-                ));
-            };
-            return executor.execute(request).await;
-        }
-        self.dispatch_executor.execute(request).await
-    }
-}
-
-fn is_process_sandbox_process(request: &ProcessExecutionRequest) -> bool {
-    request.runtime == RuntimeKind::System
-        && request.capability_id.as_str() == ironclaw_process_sandbox::PROCESS_SANDBOX_CAPABILITY_ID
-}
-
-#[async_trait]
-impl ProcessExecutor for RuntimeDispatchProcessExecutor {
-    async fn execute(
-        &self,
-        request: ProcessExecutionRequest,
-    ) -> Result<ProcessExecutionResult, ProcessExecutionError> {
-        if request.cancellation.is_cancelled() {
-            return Err(ProcessExecutionError::new("cancelled"));
-        }
-        let result = self
-            .dispatcher
-            .dispatch_json(CapabilityDispatchRequest {
-                capability_id: request.capability_id,
-                scope: request.scope,
-                estimate: request.estimate,
-                mounts: Some(request.mounts),
-                resource_reservation: request.resource_reservation,
-                input: request.input,
-            })
-            .await
-            .map_err(|error| ProcessExecutionError::new(dispatch_error_kind(&error)))?;
-        if request.cancellation.is_cancelled() {
-            return Err(ProcessExecutionError::new("cancelled"));
-        }
-        Ok(ProcessExecutionResult {
-            output: result.output,
-        })
-    }
-}
-
-fn dispatch_error_kind(error: &DispatchError) -> &'static str {
-    match error {
-        DispatchError::UnknownCapability { .. } => "unknown_capability",
-        DispatchError::UnknownProvider { .. } => "unknown_provider",
-        DispatchError::RuntimeMismatch { .. } => "runtime_mismatch",
-        DispatchError::MissingRuntimeBackend { .. } => "missing_runtime_backend",
-        DispatchError::UnsupportedRuntime { .. } => "unsupported_runtime",
-        DispatchError::Mcp { kind }
-        | DispatchError::Script { kind }
-        | DispatchError::Wasm { kind }
-        | DispatchError::FirstParty { kind } => kind.event_kind(),
     }
 }
