+++ crates/ironclaw_host_runtime/src/services/process_executor.rs
@@ -1,0 +1,278 @@
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use ironclaw_host_api::{
+    CapabilityDispatchRequest, CapabilityDispatcher, DispatchError, RuntimeKind,
+};
+use ironclaw_processes::{
+    ProcessExecutionError, ProcessExecutionRequest, ProcessExecutionResult, ProcessExecutor,
+};
+
+#[derive(Clone)]
+pub(super) struct ProcessExecutorRouter {
+    dispatch_executor: Arc<dyn ProcessExecutor>,
+    routes: Vec<ProcessExecutorRoute>,
+}
+
+impl ProcessExecutorRouter {
+    pub(super) fn new(dispatch_executor: Arc<dyn ProcessExecutor>) -> Self {
+        Self {
+            dispatch_executor,
+            routes: Vec::new(),
+        }
+    }
+
+    pub(super) fn with_optional_route(
+        mut self,
+        runtime: RuntimeKind,
+        capability_id: impl Into<String>,
+        executor: Option<Arc<dyn ProcessExecutor>>,
+        missing_error_kind: &'static str,
+    ) -> Self {
+        self.routes.push(ProcessExecutorRoute {
+            runtime,
+            capability_id: capability_id.into(),
+            executor,
+            missing_error_kind,
+        });
+        self
+    }
+}
+
+#[async_trait]
+impl ProcessExecutor for ProcessExecutorRouter {
+    async fn execute(
+        &self,
+        request: ProcessExecutionRequest,
+    ) -> Result<ProcessExecutionResult, ProcessExecutionError> {
+        if let Some(route) = self.matching_route(&request) {
+            let Some(executor) = &route.executor else {
+                return Err(ProcessExecutionError::new(route.missing_error_kind));
+            };
+            return executor.execute(request).await;
+        }
+        self.dispatch_executor.execute(request).await
+    }
+}
+
+impl ProcessExecutorRouter {
+    fn matching_route(&self, request: &ProcessExecutionRequest) -> Option<&ProcessExecutorRoute> {
+        self.routes.iter().find(|route| {
+            route.runtime == request.runtime
+                && route.capability_id.as_str() == request.capability_id.as_str()
+        })
+    }
+}
+
+#[derive(Clone)]
+struct ProcessExecutorRoute {
+    runtime: RuntimeKind,
+    capability_id: String,
+    executor: Option<Arc<dyn ProcessExecutor>>,
+    missing_error_kind: &'static str,
+}
+
+#[derive(Clone)]
+pub(super) struct RuntimeDispatchProcessExecutor {
+    dispatcher: Arc<dyn CapabilityDispatcher>,
+}
+
+impl RuntimeDispatchProcessExecutor {
+    pub(super) fn new(dispatcher: Arc<dyn CapabilityDispatcher>) -> Self {
+        Self { dispatcher }
+    }
+}
+
+#[async_trait]
+impl ProcessExecutor for RuntimeDispatchProcessExecutor {
+    async fn execute(
+        &self,
+        request: ProcessExecutionRequest,
+    ) -> Result<ProcessExecutionResult, ProcessExecutionError> {
+        if request.cancellation.is_cancelled() {
+            return Err(ProcessExecutionError::new("cancelled"));
+        }
+        let result = self
+            .dispatcher
+            .dispatch_json(CapabilityDispatchRequest {
+                capability_id: request.capability_id,
+                scope: request.scope,
+                estimate: request.estimate,
+                mounts: Some(request.mounts),
+                resource_reservation: request.resource_reservation,
+                input: request.input,
+            })
+            .await
+            .map_err(|error| ProcessExecutionError::new(dispatch_error_kind(&error)))?;
+        if request.cancellation.is_cancelled() {
+            return Err(ProcessExecutionError::new("cancelled"));
+        }
+        Ok(ProcessExecutionResult {
+            output: result.output,
+        })
+    }
+}
+
+fn dispatch_error_kind(error: &DispatchError) -> &'static str {
+    match error {
+        DispatchError::UnknownCapability { .. } => "unknown_capability",
+        DispatchError::UnknownProvider { .. } => "unknown_provider",
+        DispatchError::RuntimeMismatch { .. } => "runtime_mismatch",
+        DispatchError::MissingRuntimeBackend { .. } => "missing_runtime_backend",
+        DispatchError::UnsupportedRuntime { .. } => "unsupported_runtime",
+        DispatchError::Mcp { kind }
+        | DispatchError::Script { kind }
+        | DispatchError::Wasm { kind } => kind.event_kind(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::sync::{Arc, Mutex};
+
+    use ironclaw_host_api::{
+        AgentId, CapabilityId, ExtensionId, InvocationId, MountView, ProcessId, ProjectId,
+        ResourceEstimate, ResourceScope, TenantId, ThreadId, UserId,
+    };
+    use ironclaw_processes::ProcessCancellationToken;
+    use serde_json::json;
+
+    #[derive(Default)]
+    struct RecordingProcessExecutor {
+        label: &'static str,
+        calls: Arc<Mutex<Vec<CapabilityId>>>,
+    }
+
+    impl RecordingProcessExecutor {
+        fn new(label: &'static str) -> Self {
+            Self {
+                label,
+                calls: Arc::new(Mutex::new(Vec::new())),
+            }
+        }
+
+        fn calls(&self) -> Vec<CapabilityId> {
+            self.calls
+                .lock()
+                .unwrap_or_else(|poisoned| poisoned.into_inner())
+                .clone()
+        }
+    }
+
+    #[async_trait]
+    impl ProcessExecutor for RecordingProcessExecutor {
+        async fn execute(
+            &self,
+            request: ProcessExecutionRequest,
+        ) -> Result<ProcessExecutionResult, ProcessExecutionError> {
+            self.calls
+                .lock()
+                .unwrap_or_else(|poisoned| poisoned.into_inner())
+                .push(request.capability_id);
+            Ok(ProcessExecutionResult {
+                output: json!({ "executor": self.label }),
+            })
+        }
+    }
+
+    #[tokio::test]
+    async fn process_executor_router_sends_routed_capability_to_configured_executor() {
+        let dispatch = Arc::new(RecordingProcessExecutor::new("dispatch"));
+        let sandbox = Arc::new(RecordingProcessExecutor::new("sandbox"));
+        let router = ProcessExecutorRouter::new(dispatch.clone() as Arc<dyn ProcessExecutor>)
+            .with_optional_route(
+                RuntimeKind::System,
+                "system.process_sandbox.run",
+                Some(sandbox.clone() as Arc<dyn ProcessExecutor>),
+                "missing_process_sandbox_executor",
+            );
+
+        let result = router
+            .execute(sample_process_request(
+                "system.process_sandbox.run",
+                RuntimeKind::System,
+            ))
+            .await
+            .unwrap();
+
+        assert_eq!(result.output, json!({ "executor": "sandbox" }));
+        assert!(dispatch.calls().is_empty());
+        assert_eq!(sandbox.calls().len(), 1);
+    }
+
+    #[tokio::test]
+    async fn process_executor_router_keeps_unrouted_processes_on_dispatch_executor() {
+        let dispatch = Arc::new(RecordingProcessExecutor::new("dispatch"));
+        let sandbox = Arc::new(RecordingProcessExecutor::new("sandbox"));
+        let router = ProcessExecutorRouter::new(dispatch.clone() as Arc<dyn ProcessExecutor>)
+            .with_optional_route(
+                RuntimeKind::System,
+                "system.process_sandbox.run",
+                Some(sandbox.clone() as Arc<dyn ProcessExecutor>),
+                "missing_process_sandbox_executor",
+            );
+
+        let result = router
+            .execute(sample_process_request(
+                "demo.background",
+                RuntimeKind::Script,
+            ))
+            .await
+            .unwrap();
+
+        assert_eq!(result.output, json!({ "executor": "dispatch" }));
+        assert_eq!(dispatch.calls().len(), 1);
+        assert!(sandbox.calls().is_empty());
+    }
+
+    #[tokio::test]
+    async fn process_executor_router_fails_routed_capability_when_unconfigured() {
+        let dispatch = Arc::new(RecordingProcessExecutor::new("dispatch"));
+        let router = ProcessExecutorRouter::new(dispatch.clone() as Arc<dyn ProcessExecutor>)
+            .with_optional_route(
+                RuntimeKind::System,
+                "system.process_sandbox.run",
+                None,
+                "missing_process_sandbox_executor",
+            );
+
+        let error = router
+            .execute(sample_process_request(
+                "system.process_sandbox.run",
+                RuntimeKind::System,
+            ))
+            .await
+            .unwrap_err();
+
+        assert_eq!(error.kind, "missing_process_sandbox_executor");
+        assert!(dispatch.calls().is_empty());
+    }
+
+    fn sample_process_request(
+        capability_id: &str,
+        runtime: RuntimeKind,
+    ) -> ProcessExecutionRequest {
+        ProcessExecutionRequest {
+            process_id: ProcessId::new(),
+            invocation_id: InvocationId::new(),
+            scope: ResourceScope {
+                tenant_id: TenantId::new("tenant").unwrap(),
+                user_id: UserId::new("user").unwrap(),
+                agent_id: Some(AgentId::new("agent").unwrap()),
+                project_id: Some(ProjectId::new("project").unwrap()),
+                mission_id: None,
+                thread_id: Some(ThreadId::new("thread").unwrap()),
+                invocation_id: InvocationId::new(),
+            },
+            extension_id: ExtensionId::new("system").unwrap(),
+            capability_id: CapabilityId::new(capability_id).unwrap(),
+            runtime,
+            estimate: ResourceEstimate::default(),
+            mounts: MountView::default(),
+            resource_reservation: None,
+            input: json!({}),
+            cancellation: ProcessCancellationToken::new(),
+        }
+    }
+}
